Fix npm cache server

+++ src/runtime/assets-processer/ts/spec/proxy-cache/test-proxy-cache.ts
@@ -7,22 +7,25 @@ import fs from 'fs-extra';
 import {config, runServer} from '@wfh/plink';
 const isWin = os.platform() === 'win32';
 
-export async function runNpmInstall() {
-  const serverCache = Path.resolve(config().destDir, 'npm-server-cache');
-  if (fs.existsSync(serverCache))
-    await fs.remove(serverCache);
+async function runNpmInstall(clean: boolean) {
+  if (clean) {
+    const serverCache = Path.resolve(config().destDir, 'npm-server-cache');
+    if (fs.existsSync(serverCache))
+      await fs.remove(serverCache);
+  }
 
   config.change(setting => {
     setting['@wfh/assets-processer'].npmRegistryCacheServer = {
       path: '/npm',
       cacheDir: setting.destDir + '/npm-server-cache',
-      registry: 'https://registry.npm.taobao.org/'
+      // registry: 'https://registry.npm.taobao.org/'
+      registry: 'https://registry.npmjs.org/'
     };
   });
 
   const {started, shutdown} = runServer();
   await started;
-  await rx.of('node_modules', 'npm-cache').pipe(
+  await rx.of('node_modules', 'package-lock.json', 'npm-cache').pipe(
     op.mergeMap(dir => {
       const target = Path.resolve(__dirname, dir);
       if (fs.existsSync(target)) {
@@ -32,9 +35,9 @@ export async function runNpmInstall() {
     })
   ).toPromise();
 
-  const cp = spawn(isWin ? 'npm.cmd' : 'npm', ['install', 'glob'], {
+  const cp = spawn(isWin ? 'npm.cmd' : 'npm', ['install', '--ddd'], {
     cwd: __dirname, stdio: 'inherit',
-    timeout: 45000
+    timeout: 65000
   });
 
   const error$ = rx.fromEventPattern<Error>(
@@ -45,7 +48,7 @@ export async function runNpmInstall() {
 
   const exit$ = rx.fromEventPattern(h => cp.on('exit', h), h => cp.off('exit', h));
 
-  return rx.merge(error$, exit$, rx.timer(46000)).pipe(
+  return rx.merge(error$, exit$, rx.timer(3 * 60000)).pipe(
     op.take(1),
     op.catchError(err => {
       // eslint-disable-next-line no-console
@@ -62,5 +65,54 @@ export async function runNpmInstall() {
     op.tap(() => process.exit(0))
   ).toPromise();
 
+}
+
+export async function npmInstall() {
+  await rx.of('node_modules', 'package-lock.json', 'npm-cache').pipe(
+    op.mergeMap(dir => {
+      const target = Path.resolve(__dirname, dir);
+      if (fs.existsSync(target)) {
+        return fs.remove(target);
+      }
+      return rx.EMPTY;
+    })
+  ).toPromise();
+
+  const cp = spawn(isWin ? 'npm.cmd' : 'npm', ['install', '--ddd'], {
+    cwd: __dirname, stdio: 'inherit',
+    timeout: 65000
+  });
+
+  const error$ = rx.fromEventPattern<Error>(
+    h => cp.on('error', h), h => cp.off('error', h)
+  ).pipe(
+    op.switchMap(err => rx.throwError(err))
+  );
+
+  const exit$ = rx.fromEventPattern(h => cp.on('exit', h), h => cp.off('exit', h));
+
+  return rx.merge(error$, exit$, rx.timer(3 * 60000)).pipe(
+    op.take(1),
+    op.catchError(err => {
+      // eslint-disable-next-line no-console
+      console.log('test npm install failed:', err);
+      return rx.of(1);
+    }),
+    op.tap(() => {
+      // eslint-disable-next-line no-console
+      console.log('kill: ', cp.kill('SIGINT'));
+      // await shutdown();
+      // eslint-disable-next-line no-console
+      console.log('------- shutdown http server done ---------');
+    })
+  ).toPromise();
+}
+
+export function test() {
+  return runNpmInstall(false);
+}
+
+export function cleanTest() {
+  return runNpmInstall(true);
 }
